+++ WalkFileTree.c
@@ -21,60 +21,45 @@
 #include "stringutil.h"
 #include "edierror.h"
 
-extern int string_matchFilename(char *string,char *pat);
-int __ftw(char *path,int depth);
-
-/* to be reset before every search - indicates minimum 1 file found */
-int _ftwflg;
-
-static char *_searchpattern;
-static int  _mode;
-static int  (*_func)();
-
 /*--------------------------------------------------------------------------
- * __ftw()
- * uses #include ftw.h
+ * _ftw()
+ * perform a file tree walk scanning files recursively given a file name pattern to match
+ * and executes the given callback on each matching file.
  */
 int _ftw(
-	char *path,				/* Root where the file tree walk begins		*/
-	FTWFUNC func,			/* execute for every file found, until func returns != 0 or ... */
-	int depth,				/* the depth of the maximum folder recursion level is reached*/
-	char *pattern, 			/* pattern for file name matching */
-	int mode				/* file attributes of the files searched */
-)
-{	int ret;
-
-	_searchpattern = pattern;
-	_mode = mode;
-	_func = func;
-	ret = __ftw(path,depth);
-	return ret;
-}
-
-/*--------------------------------------------------------------------------
- * __ftw()
- */
-static int __ftw(char *path,int depth)
-{	struct _finddata_t  *	pdta;
+	const char *path,				// Root where the file tree walk begins
+	FTWFUNC func,					// execute for every file found, until func returns != 0 or ... 
+	int depth,						// the depth of the maximum folder recursion level is reached
+	const char *searchPattern, 		// pattern for file name matching. E.g. something such as "*.c" or "dir\*.c". In the later case the pattern is matched
+									// also recursively for sub-directories.
+	int mode						// file attributes of the files searched
+) {
+	struct _finddata_t  *	pdta;
 	char 				*	target;
 	intptr_t				fhandle;
 	int						nFiles = 0;
-	register int			i;
+	int						i;
+	const char*				pThisPattern = searchPattern;
+	const char*				pNextPattern = searchPattern;
 
 	if (--depth < 0) {
 		return 0;
 	}
+	char* pPatternCopy = _strdup(searchPattern);
 	i = 0;
 	pdta = malloc(sizeof *pdta);
 	target = malloc(strlen(path) + 256);
 	string_concatPathAndFilename(target,path,"*");
+	char* pToken = strtok(pPatternCopy, "/\\");
+	if (strcmp(pToken, searchPattern) != 0) {
+		pThisPattern = pToken;
+		pNextPattern = pPatternCopy+strlen(pThisPattern)+1;
+	}
 	if ((fhandle = _findfirst(target, pdta)) >= 0) {
 		do {
 			if (progress_cancelMonitor(0)) {
-				_findclose(fhandle);
-				free(pdta);
-				free(target);
-				return 1;
+				i = 1;
+				goto done;
 			}
 
 			if (strcmp(pdta->name, ".") == 0 || strcmp(pdta->name, "..") == 0) {
@@ -82,37 +67,32 @@ static int __ftw(char *path,int depth)
 			}
 			nFiles++;
 			if (pdta->attrib == 0 || 
-			   (pdta->attrib & _mode) == pdta->attrib) {
-				if ((pdta->attrib & _A_SUBDIR) || 
-					 string_matchFilename(pdta->name,_searchpattern)) {
-					_ftwflg++;
+			   (pdta->attrib & mode) == pdta->attrib) {
+				if ((pThisPattern == searchPattern && (pdta->attrib & _A_SUBDIR)) || 
+					 string_matchFilename(pdta->name,pThisPattern)) {
 					string_concatPathAndFilename(target,path,pdta->name);
-					i = (*_func)(target,pdta);
+					i = (*func)(target,pdta);
 					if (i) {
-						free(pdta);
-						free(target);
-						_findclose(fhandle);
-						return i;
+						goto done;
 					}
 				} else if (nFiles % 10 == 9) {
 					progress_stepIndicator();
 				}
 			}
 
-			if (depth > 0 && (pdta->attrib & _A_SUBDIR) != 0) {
+			if (depth > 0 && (pdta->attrib & _A_SUBDIR) != 0 && (pThisPattern == searchPattern || string_matchFilename(pdta->name, pThisPattern))) {
 				string_concatPathAndFilename(target,path,pdta->name);
-				i = __ftw(target,depth);
+				i = _ftw(target, func, depth, pNextPattern,  mode);
 				if (i != 0) {
-					free(pdta);
-					free(target);
-					return i;
+					goto done;
 				}
 			}
 
 		} while (!_findnext(fhandle, pdta));
-		_findclose(fhandle);
 	}
-
+done:
+	_findclose(fhandle);
+	free(pPatternCopy);
 	free(pdta);
 	free(target);
 	return i;